refactor: Make cache expiry a Cacher implementation detail
